kato-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.74.0
protobuf==6.32.0
pyyaml==6.0.1
python-multipart==0.0.6
httpx==0.25.1
prometheus-client==0.19.0